homepage/mysite/requirements.txt
aldryn-background-image==0.1.1
aldryn-boilerplates==0.7.4
aldryn-bootstrap3==1.1.2
aldryn-categories==1.0.3
aldryn-common==1.0.4
aldryn-search==0.3.0
aldryn-style==1.1.1
aldryn-translation-tools==0.2.1
cmsplugin-filer==1.1.3
dj-database-url==0.4.1
Django==1.8.16
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.1
django-durationfield==0.5.2
django-filer==1.2.5
django-formtools==1.0
django-haystack==2.5.1
django-mptt==0.8.6
django-parler==1.6.5
django-polymorphic==0.8.1
django-sekizai==0.10.0
Django-Select2==4.3.2
django-sortedm2m==1.3.2
django-spurl==0.6.4
django-standard-form==1.1.1
django-treebeard==4.0.1
djangocms-admin-style==1.2.6
djangocms-attributes-field==0.1.2
djangocms-column==1.7.0
djangocms-googlemap==0.5.2
djangocms-installer==0.9.1
djangocms-link==2.0.2
djangocms-snippet==1.9.1
djangocms-style==1.7.0
djangocms-text-ckeditor==3.3.1
djangocms-video==2.0.2
djangorestframework==3.5.3
djoser==0.5.1
easy-thumbnails==2.3
html5lib==0.9999999
lxml==3.6.4
Pillow==3.4.2
python-slugify==1.2.0
pytz==2016.7
requests==2.11.1
six==1.10.0
tzlocal==1.3
Unidecode==0.4.19
URLObject==2.4.2
YURL==0.13
